src/main/java/com/cheese/demo/discount/DiscountServiceImpl.java
package com.cheese.demo.discount;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
public class DiscountServiceImpl implements DiscountService {
    private final DiscountRepository discountRepository;

    @Autowired
    public DiscountServiceImpl(DiscountRepository discountRepository) {
        this.discountRepository = discountRepository;
    }

    @Override
    public Discount create(DiscountDto.Creation dto) {
        return discountRepository.save(dto.toEntity());
    }
}